requirements.txt
click==8.0.3
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
beautifulsoup4==4.12.2
requests==2.28.1  # Updated version
urllib3==1.26